+++ aldrin/src/client.rs
@@ -374,7 +374,7 @@ where
             Message::BusListenerCurrentFinished(msg) => {
                 self.msg_bus_listener_current_finished(msg)?
             }
-            Message::AbortFunctionCall(_) => todo!(),
+            Message::AbortFunctionCall(_) => {}
 
             Message::Connect(_)
             | Message::ConnectReply(_)